+++ src/Method/SetChatAdministratorCustomTitle.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method;
+
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+
+/**
+ * @see https://core.telegram.org/bots/api#setchatadministratorcustomtitle
+ */
+final readonly class SetChatAdministratorCustomTitle implements TelegramRequestWithResultPreparingInterface
+{
+    public function __construct(
+        private int|string $chatId,
+        private int $userId,
+        private string $customTitle
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'setChatAdministratorCustomTitle';
+    }
+
+    public function getData(): array
+    {
+        return [
+            'chat_id' => $this->chatId,
+            'user_id' => $this->userId,
+            'custom_title' => $this->customTitle,
+        ];
+    }
+
+    public function prepareResult(mixed $result): true
+    {
+        return true;
+    }
+}